requirements.txt
python-dotenv
google.generativeai
PyPDF2
jinja2
python-docx